# Save to cache info about broken, too short etc. images and videos