Fixes https://github.com/qarmin/czkawka/issues/452

It really helps to decrease amount of time spend on hashing invalid/broken/too short video files.
With images, speedup is a lot of smaller(but still visible, especially when trying to check big amount of broken files)

Tested 186 videos

Old version:
First scan: 75s
Second scan: 16s

New version:
First scan: 75s
Second scan: 2s